refactor formatters
